data/scripts/requirements.txt
# Python dependencies for knowledge base building
# Install with: pip3 install -r requirements.txt

python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.21.0
